loans/requirement.txt
asgiref==3.8.1
django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
